Enforce the RUF rules the review cited, and make CLI tests host-independent
Both halves of this came out of a disagreement worth recording.

**The lint gate.** A review finding said `ruff check .` fails with
RUF022/RUF012/RUF059; I replied that those rules are not enabled here
and the gate was clean. Re-checked, we were each half right, and the
ambiguity was the real defect: with only `line-length` and
`target-version` set, ruff falls back to its default `E4,E7,E9,F`, so
`ruff check .` genuinely passes — while `--select RUF` finds 25 real
violations at exactly the lines the review named.

A config that silent lets two readers reach opposite conclusions about
whether the gate is green, so the selection is now explicit and the
violations are fixed:

- RUF012 on 19 mutable class attributes, 18 of them Textual `BINDINGS`.
  `ClassVar[list[BindingType]]` is the annotation Textual documents.
- RUF022 (`__all__` unsorted), RUF046 (`int(round(...))`, where `round`
  already returns an int), RUF059 (three unused unpacked variables).

Deliberately left out of the selection: RUF001, because the typographic
apostrophe in `utils/text.py` is intentional — it makes `_WORD` match
"don't" with either quote character — and RUF100, whose 65 hits are an
artifact of narrow selection making existing noqa codes look unused.

**Host-dependent tests.** `test_doctor_reports_problems_with_a_nonzero_exit`
asserted exit code 1 with a docstring conceding "this environment has no
whisper.cpp"; `test_models_whisper_lists_availability` required `small.en`
to be absent; `test_models_ollama_reports_a_missing_daemon` probed the real
localhost:11434 and required nothing to answer. All three passed here only
because the machine lacks the software, and would have failed on a Mac with
Lectern's own dependencies installed.

They now stub the probe and assert on the command's behaviour, and each
gained the opposite case the old version could never reach: doctor exiting
zero when checks pass, and Ollama listing models when it is running.

One of the replacements was initially weak in the way this file already
warns about — "installed" also matches inside "not installed" — so it now
parses the rendered rows and asserts the two states distinguish, verified
by pointing the installed model at a path that does not exist and watching
it fail.

202 tests pass; `ruff check .` clean under the wider selection.

+++ src/lectern/utils/audio_utils.py
@@ -37,7 +37,7 @@ def resample(audio: np.ndarray, source_rate: int, target_rate: int = TARGET_SAMP
         audio = _lowpass(audio, cutoff=target_rate / 2.0, sample_rate=source_rate)
 
     duration = audio.size / source_rate
-    out_samples = int(round(duration * target_rate))
+    out_samples = round(duration * target_rate)
     if out_samples <= 0:
         return np.zeros(0, dtype=np.float32)
